+++ src/app/private/SignOutButton.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { Button } from '@/common/ui/Button/Button'
-import { createClient } from '@/common/data/supabase/client'
+import { createClient } from '@/common/services/supabase/client'
 import { useRouter } from 'next/navigation'
 import { useState } from 'react'
